backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.36.2
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.27.2